\name{xmapbridge-internal}
\title{Internal xmapbridge objects}
\alias{.get.xmap.cache.dir}
\alias{.make.xmap.child.file}
\alias{.make.xmap.project.folder}
\alias{.xmap.env}
\alias{.xmap.graph.valid}
\alias{.xmap.project.valid}
\alias{.xmap.read.file.info}
\alias{.xmap.write.file.info}
\description{Internal xmapbridge objects.}
\details{These are not to be called by the user.}
\keyword{internal}