\name{xmapbridge-package}
\alias{xmapbridge-package}
\alias{xmapbridge}
\docType{package}
\title{
  Export plotting files to the xmapBridge for visualisation in X:Map
}
\description{
  The XMapBridge is a Java application which allows you to plot graphs in the X:Map genome browser.  This package exports plotting files in a suitable format for the XMapBridge to read.
}
\details{
  \tabular{ll}{
    Package: \tab xmapbridge\cr
    Type: \tab Package\cr
    Version: \tab 1.0\cr
    Date: \tab 2008-08-11\cr
    License: \tab LGPL-v3\cr
  }
}
\author{
  Tim Yates and Crispin J Miller\cr
  Maintainer: Tim Yates <tyates@picr.man.ac.uk>
}
\references{
  Yates T, Okoniewski MJ, Miller CJ. X:Map: annotation and visualization of genome structure for Affymetrix exon array analysis. Nucleic Acids Res. 2008 Jan;36(Database issue):D780-6. Epub 2007 Oct 11.\cr
  \url{http://nar.oxfordjournals.org/cgi/content/full/gkm779v1}
}
\keyword{ package }
\seealso{
  \code{\link{xmap.plot}}\cr
  \code{\link{xmap.points}}\cr
  \code{\link{xmap.col}}\cr
  \code{\link{xmap.project.new}}\cr
  \code{\link{xmap.project.list}}\cr
  \code{\link{xmap.project.load}}\cr
  \code{\link{xmap.project.save}}\cr
  \code{\link{xmap.project.delete}}\cr
  \code{\link{xmap.graph.new}}\cr
  \code{\link{xmap.graph.list}}\cr
  \code{\link{xmap.graph.load}}\cr
  \code{\link{xmap.graph.save}}\cr
  \code{\link{xmap.graph.delete}}\cr
  \code{\link{xmap.plot.new}}\cr
  \code{\link{xmap.plot.list}}\cr
  \code{\link{xmap.plot.load}}\cr
  \code{\link{xmap.plot.save}}\cr
  \code{\link{xmap.plot.delete}}
}
